requirements.txt
absl-py==2.1.0
asttokens==2.4.1
astunparse==1.6.3
bleach==6.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.0
fsspec==2024.10.0
gast==0.6.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.68.0
h5py==3.12.1
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
kaggle==1.6.17
kagglehub==0.3.4
keras==2.15.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
ml-dtypes==0.2.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==4.25.5
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.14.1
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
text-unidecode==1.3
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.14.1
